requirements.txt
flask==1.1.2
numpy==1.19.5
beautifulsoup4==4.6.3
requests==2.20.1
scipy
pandas==1.3.1
lxml==4.5.2
line-bot-sdk==1.19.0
datetime 
matplotlib
gspread==4.0.1
oauth2client==4.1.3
pyimgur
mpl_finance
arrow
